requirements.txt
SystemExit: 2


Django==4.0.6
Django_Verify_Email==2.0.3
Jinja2==3.1.2
MarkupSafe==2.1.1
Pillow==9.2.0
PyJWT==2.4.0
asgiref==3.5.2
backports.zoneinfo==0.2.1
certifi==2022.6.15
cffi==1.15.1
charset_normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.4
defusedxml==0.7.1
dj_rest_auth==2.2.5
django_cors_headers==3.13.0
django_simple_email_confirmation==0.70
django_templated_mail==1.1.1
djangorestframework==3.13.1
djangorestframework_api_key==2.2.0
djangorestframework_jwt==1.11.0
djangorestframework_simplejwt==4.8.0
djoser==2.1.0
drf_writable_nested==0.7.0
gunicorn==20.1.0
idna==3.3
itypes==1.2.0
oauthlib==3.2.0
pip==22.2.2
pycparser==2.21
python3_openid==3.2.0
pytz==2022.1
requests==2.28.1
requests_oauthlib==1.3.1
setuptools==63.3.0
six==1.16.0
social_auth_app_django==4.0.0
social_auth_core==4.3.0
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.12
wheel==0.37.1
whitenoise==6.2.0